+++ workspaces/src/types/account.rs
@@ -18,14 +18,15 @@ use crate::result::{Execution, ExecutionFinalResult, Result};
 /// network, such as creating transactions and calling into contract functions.
 #[derive(Clone)]
 pub struct Account {
-    pub(crate) id: AccountId,
-    pub(crate) signer: InMemorySigner,
+    signer: InMemorySigner,
     worker: Worker<dyn Network>,
 }
 
 impl fmt::Debug for Account {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("Account").field("id", &self.id).finish()
+        f.debug_struct("Account")
+            .field("id", &self.signer.account_id)
+            .finish()
     }
 }
 
@@ -36,8 +37,7 @@ impl Account {
         worker: &Worker<impl Network + 'static>,
     ) -> Result<Self> {
         let signer = InMemorySigner::from_file(path.as_ref())?;
-        let id = signer.account_id.clone();
-        Ok(Self::new(id, signer, worker.clone().coerce()))
+        Ok(Self::new(signer, worker.clone().coerce()))
     }
 
     /// Create an [`Account`] object from an [`AccountId`] and [`SecretKey`].
@@ -47,22 +47,23 @@ impl Account {
         worker: &Worker<impl Network + 'static>,
     ) -> Self {
         Self {
-            id: id.clone(),
             signer: InMemorySigner::from_secret_key(id, sk),
             worker: worker.clone().coerce(),
         }
     }
 
-    pub(crate) fn new(id: AccountId, signer: InMemorySigner, worker: Worker<dyn Network>) -> Self {
-        Self { id, signer, worker }
+    pub(crate) fn new(signer: InMemorySigner, worker: Worker<dyn Network>) -> Self {
+        Self { signer, worker }
     }
 
     /// Grab the current account identifier
     pub fn id(&self) -> &AccountId {
-        &self.id
+        &self.signer.account_id
     }
 
-    pub(crate) fn signer(&self) -> &InMemorySigner {
+    /// Grab the signer of the account. This signer is used to sign all transactions
+    /// sent to the network.
+    pub fn signer(&self) -> &InMemorySigner {
         &self.signer
     }
 
@@ -70,13 +71,9 @@ impl Account {
     /// a [`CallTransaction`] object that we will make use to populate the
     /// rest of the call details. Note that the current [`Account`]'s secret
     /// key is used as the signer of the transaction.
-    pub fn call<'a, 'b>(
-        &'a self,
-        contract_id: &AccountId,
-        function: &'b str,
-    ) -> CallTransaction<'a> {
+    pub fn call(&self, contract_id: &AccountId, function: &str) -> CallTransaction {
         CallTransaction::new(
-            &self.worker,
+            self.worker.clone(),
             contract_id.to_owned(),
             self.signer.clone(),
             function,
@@ -105,13 +102,13 @@ impl Account {
     /// transaction. The beneficiary will receive the funds of the account deleted
     pub async fn delete_account(self, beneficiary_id: &AccountId) -> Result<ExecutionFinalResult> {
         self.worker
-            .delete_account(&self.id, &self.signer, beneficiary_id)
+            .delete_account(self.id(), &self.signer, beneficiary_id)
             .await
     }
 
     /// Views the current account's details such as balance and storage usage.
     pub fn view_account(&self) -> Query<'_, ViewAccount> {
-        self.worker.view_account(&self.id)
+        self.worker.view_account(self.id())
     }
 
     /// Views the current accounts's access key, given the [`PublicKey`] associated to it.
@@ -133,7 +130,7 @@ impl Account {
         Query::new(
             self.worker.client(),
             ViewAccessKeyList {
-                account_id: self.id.clone(),
+                account_id: self.id().clone(),
             },
         )
     }
@@ -163,11 +160,7 @@ impl Account {
             .await?;
 
         Ok(Execution {
-            result: Contract::new(
-                self.id().clone(),
-                self.signer().clone(),
-                self.worker.clone(),
-            ),
+            result: Contract::new(self.signer().clone(), self.worker.clone()),
             details: ExecutionFinalResult::from_view(outcome),
         })
     }
@@ -179,7 +172,7 @@ impl Account {
     /// network.
     pub fn batch(&self, contract_id: &AccountId) -> Transaction {
         Transaction::new(
-            self.worker.client(),
+            self.worker.clone(),
             self.signer().clone(),
             contract_id.clone(),
         )
@@ -187,15 +180,10 @@ impl Account {
 
     /// Store the credentials of this account locally in the directory provided.
     pub async fn store_credentials(&self, save_dir: impl AsRef<Path>) -> Result<()> {
-        let savepath = save_dir.as_ref().to_path_buf();
-        std::fs::create_dir_all(save_dir).map_err(|e| ErrorKind::Io.custom(e))?;
-
-        let mut savepath = savepath.join(self.id.to_string());
-        savepath.set_extension("json");
-
-        crate::rpc::tool::write_cred_to_file(&savepath, &self.id, &self.secret_key().0);
-
-        Ok(())
+        let savepath = save_dir.as_ref();
+        std::fs::create_dir_all(&save_dir).map_err(|e| ErrorKind::Io.custom(e))?;
+        let savepath = savepath.join(format!("{}.json", self.id()));
+        crate::rpc::tool::write_cred_to_file(&savepath, self.id(), &self.secret_key().0)
     }
 
     /// Get the keys of this account. The public key can be retrieved from the secret key.
@@ -221,7 +209,7 @@ pub struct Contract {
 impl fmt::Debug for Contract {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_struct("Contract")
-            .field("id", &self.account.id)
+            .field("id", self.account.id())
             .finish()
     }
 }
@@ -236,9 +224,9 @@ impl Contract {
         Self::account(Account::from_secret_key(id, sk, worker))
     }
 
-    pub(crate) fn new(id: AccountId, signer: InMemorySigner, worker: Worker<dyn Network>) -> Self {
+    pub(crate) fn new(signer: InMemorySigner, worker: Worker<dyn Network>) -> Self {
         Self {
-            account: Account::new(id, signer, worker),
+            account: Account::new(signer, worker),
         }
     }
 
@@ -248,7 +236,7 @@ impl Contract {
 
     /// Grab the current contract's account identifier
     pub fn id(&self) -> &AccountId {
-        &self.account.id
+        self.account.id()
     }
 
     /// Treat this [`Contract`] object as an [`Account`] type. This does nothing
@@ -265,7 +253,9 @@ impl Contract {
         &mut self.account
     }
 
-    pub(crate) fn signer(&self) -> &InMemorySigner {
+    /// Grab the signer of the account. This signer is used to sign all transactions
+    /// sent to the network.
+    pub fn signer(&self) -> &InMemorySigner {
         self.account.signer()
     }
 
@@ -275,7 +265,7 @@ impl Contract {
     ///
     /// If we want to make use of the contract's secret key as a signer to call
     /// into another contract, use `contract.as_account().call` instead.
-    pub fn call(&self, function: &str) -> CallTransaction<'_> {
+    pub fn call(&self, function: &str) -> CallTransaction {
         self.account.call(self.id(), function)
     }
 
@@ -314,7 +304,7 @@ impl Contract {
     }
 
     /// Deletes the current contract, and returns the execution details of this
-    /// transaction. The beneciary will receive the funds of the account deleted
+    /// transaction. The beneficiary will receive the funds of the account deleted
     pub async fn delete_contract(self, beneficiary_id: &AccountId) -> Result<ExecutionFinalResult> {
         self.account.delete_account(beneficiary_id).await
     }
@@ -347,7 +337,7 @@ impl AccountDetails {
         AccountView {
             amount: self.balance,
             locked: self.locked,
-            // unwrap guranteed to succeed unless CryptoHash impls have changed in near_primitives.
+            // unwrap guaranteed to succeed unless CryptoHash impls have changed in near_primitives.
             code_hash: near_primitives::hash::CryptoHash(self.code_hash.0),
             storage_usage: self.storage_usage,
             storage_paid_at: self.storage_paid_at,